Fix errors in slice cardinality when extracting a contains rule (#139)
* Check snapshot for slice cardinality

If no cardinality information is present on a slice element in the
differential, check the snapshot to see if the cardinality information
exists there. If cardinality information is not in either location, use
default cardinality values: 0 for min, and the sliced element's max for
max.

* Set both min and max for cardinality on a contains rule

When extracting a contains rule, the cardinality rule should include
both a min and a max. Check the differential, the snapshot, and the
sliced element, in that order, for cardinality information.