Keep paths for rules that need paths
Some rule types must have a non-empty path in FSH. When using context
paths, make sure that the path for those rules are not empty.

Rules that have empty paths or paths to the root element have special
handling when building the array of their path parts. An empty path has
an empty array of parts, and the path to the root element has an array
with "." as the only entry. Other paths are handled by splitting on the
"." character.