Make associations in contains rule based on end of path
The path of a contains rule may be changed by optimizers. Because the
paths on the associated card and flag rules do not get changed, the
associated rules' paths may not start with the contains rule's path. It
is sufficient to check that the associated rule's path ends with the
contained item's name in brackets.

When resetting context in the rule path contexts optimizer, only add the
current rule's path if it is not empty.