Use Instance meta.profile only when there is exactly one
When constructing an Instance, profiles may be present on the
meta.profile property. If meta.profile contains exactly one profile, and
that profile can be found, use that profile as the InstanceOf for the
Instance and remove the rule that sets meta.profile. If that profile is
not found, log a warning and use the Instance's resourceType property as
the InstanceOf. If there are multiple profiles in meta.profile, do not
attempt to find any of them. Instead, use the resourceType as
InstanceOf. If there are no profiles, use the resourceType as
InstanceOf.

Instances may be constructed by InstanceProcessor or by
ConstructInlineInstanceOptimizer. Test cases for these classes are
updated so that entries in meta.profile will resolve when required by
the test.