Construct inline instances before adding reference keywords
By constructing inline instances first, inline instances contained
within Bundle resources are more likely to be able to take advantage of
the Reference keyword. This is because adding the Reference keyword for
an element's value depends on checking the element's type as defined by
the StructureDefinition for the instance's type. In a Bundle, not much
type information is available for contained resources. But once the
contained resource is constructed as its own inline instance, more type
information is available.