Add option to indent rules (#138)
* Add Rule Path Contexts optimizer

This new optimizer attempts to make use of context paths in rules. If a
rule's path can be partially or entirely represented by preceding
rule's path context, set an indent on the current rule and shorten its
path to only contain the parts that come after the established context.

This new optimizer needs to run after all other optimizers, including
the array indexing optimizer. Several optimizers have their ordering
specifications changed to enforce this.

The toFSH methods on rules are not yet updated.

* Rules produce indented FSH

Rules that can be indented now add indentation to their FSH
representation. The amount of space per indentation level is specified
by a shared constant.

* Track enabled optimizers when loading

When loading optimizers, some optional optimizers may not be enabled.
Other optimizers may reference them in their runBefore or runAfter
properties. If an optimizer that will run is configured to run before
the disabled optimizer, this does not represent an error case. However,
if an optimizer that will run is configured to run after a disabled
optimizer, this represents an error case. Show error messages specific
to whether an optimizer listed as an adjacency isn't found or is turned
off.

* Simplify path contexts on Logicals and Resources

AddElementRule now implements ExportableRule so that Logicals and
Resources can be processed by the optimizer for simplifying path
contexts.

Fix tests for simplifying path contexts so that the original rules are
cloned before being modified by the optimizer.

* Set default indent for ExportableRule implementations

Rule classes that implement ExportableRule and use the indent property
now assign the value of a default of 0. This helps avoid having to check
for undefined.

Change name of OptimizerPlugin's toggle function to "isEnabled", as this
more strongly hints that it will return a boolean that indicates whether
the function is enabled or not.

* Move optimizer enable check out of loadOptimizers to getResources

All optimizers present in the file system are returned from
loadOptimizers, regardless of whether or not they are enabled. The check
for whether an optimizer is enabled is moved to immediately before the
optimizer will be run. Errors are not logged when a preceding optimizer
is not enabled. Essentially, the principle to keep in mind is that no
optimizer is ever required to be run. As long as the correct ordering is
followed, output should remain valid after running all enabled
optimizers. Tests regarding enabling optimizers are moved to the
Processing module test suite.

Rename optimizer options to processing options, as they may be used for
other parts of processing.

* Remove preceding optimizer requirement for SimplifyInstanceNameOptimizer

It is not necessary for SimplifyInstanceNameOptimizer to run before
SimplifyArrayIndexingOptimizer, since the former operates on instance
names and the latter operates on array indices in rule paths.

* Remove preceding mapping optimizer requirement

It is not necessary for the rule path contexts optimizer to run after
the mapping names optimizer, since the former operates on rule paths and
indents, and the latter operates on mapping names.

* Keep paths for rules that need paths

Some rule types must have a non-empty path in FSH. When using context
paths, make sure that the path for those rules are not empty.

Rules that have empty paths or paths to the root element have special
handling when building the array of their path parts. An empty path has
an empty array of parts, and the path to the root element has an array
with "." as the only entry. Other paths are handled by splitting on the
"." character.

* Make associations in contains rule based on end of path

The path of a contains rule may be changed by optimizers. Because the
paths on the associated card and flag rules do not get changed, the
associated rules' paths may not start with the contains rule's path. It
is sufficient to check that the associated rule's path ends with the
contained item's name in brackets.

When resetting context in the rule path contexts optimizer, only add the
current rule's path if it is not empty.

* Obeys rule on root element outputs . for path

An obeys rule on the root element of a structure definition can be
specified with no path or with a path of ".". However, using no path can
cause problems when the path context optimizer modifies rules.
Therefore, always output the "." path on obeys rules.

* Construct inline instances before adding reference keywords

By constructing inline instances first, inline instances contained
within Bundle resources are more likely to be able to take advantage of
the Reference keyword. This is because adding the Reference keyword for
an element's value depends on checking the element's type as defined by
the StructureDefinition for the instance's type. In a Bundle, not much
type information is available for contained resources. But once the
contained resource is constructed as its own inline instance, more type
information is available.

* Simplify obeys rules with . paths

Add an optimizer to change obeys rules with "." as their path to have an
empty path instead. This is a closer match to the authoring style
recommended by the FSH documentation. Because this changes paths on
rules, the rule path contexts optimizer must run after this new
optimizer.

* Add Reference Keywords after simplifying Instance names

By simplifying Instance names first, the Reference keywords will be able
to refer to the simplified name of the Instance. Then later on, when
Reference assignments are resolved, the Instance will be found with the
simplified name.