Add option for meta.profile behavior in Instances
When building an Instance, the value to use for InstanceOf can depend on
the meta.profile and resourceType attributes of the input JSON. The
strategy to use depends on a command-line option provided by the user.

"only-one" strategy: if there is exactly one entry in the meta.profile
array, try to use it as InstanceOf. If it is found when fishing, do not
add an assignment rule to set it in the Instance's meta.profile array.
If it is not found when fishing, log a warning and use resourceType as
InstanceOf. If there is not exactly one entry in the meta.profile array,
use the resourceType as InstanceOf. This is the default strategy used
when the command line option is not provided.

"first" strategy: If there is at least one entry in the meta.profile
array, try to use it as InstanceOf. If it is found when fishing, do not
add an assignment rule to set it in the Instance's meta.profile array.
If it is not found when fishing, log a warning and use resourceType as
InstanceOf. If there are no entries in the meta.profile array, use the
resourceType as InstanceOf.

"none" strategy: Ignore meta.profile when determining InstanceOf. Always
use resourceType as InstanceOf.

ProcessingOptions are now passed to InstanceProcessor and optimizer
functions. Currently, ConstructInlineInstanceOptimizer is the only
optimizer that uses the options in its optimize implementation.

Add some fields to the definition of the ProcessingOptions type to
improve hinting.