Log better errors and handle cases of malformed extensions
Certain patterns of definining extensions are processed by GoFSH
to produce invalid FSH, because there is no way to create FSH which
results in the original JSON due to some stylistic choices FSH makes.
This updates GoFSH to log an error in those situations, but generate
valid FSH, even though we know it will result in FSH that generates
technically different JSON. But the JSON has the same meaning.