Add option to generate FSH without Aliases
When the "--no-alias" flag is specified, FSH is produced without any
aliases. URLs that do not resolve to a name will be used directly rather
than being replaced by an alias.

Due to some normally helpful features of commander, do not add
command-line options directly to the Command instance as properties.
Instead, retrieve the options using the provided opts() method. This is
done to avoid colliding with the existing "alias" property on the
Command instance.