Updates to support for loading existing aliases (mint-thompson)
Squashed from:

Find aliases using FSH parser
In order to detect aliases accurately, use the FSH grammar defined by
SUSHI.
Remove ExportableAlias.fromFSH
Duplicate alias is only an error if it has a different url
Adding an alias with the same name as an existing alias but a different
url is an error. Adding an alias with the same name and url as an
existing alias is not an error. In both of these cases, the alias is not
added to the Package.
Fix bug that was preventing the alias-file flag from working.
Fix condition for detecting duplicate aliases