Add element rule may have contentReference
An ElementDefinition may have no types, and instead have a
contentReference. In this case, set the rule's contentReference and do
not emit a warning. If the ElementDefinition has at least one type and
also has a contentReference, however, that is not a valid
ElementDefinition. In this case, emit a warning and ignore the
contentReference.