Support for extensions on primitives (such as type._targetProfile.extension) (#214)
* Broken test for extensions on type.targetProfile

* Update to SUSHI 2.10.0 and Fix Instance Name/Id Problems

Update from SUSHI 2.8.0 to 2.9.0. This caused problems related to a fix in SUSHI 2.9.0 that made it so an instance's id getter would consult assignment rules for id overrides. This required:
* setting inline instances via name rather than id
* updating the name optimizer to fixup changed names in inline instane assignments

There was also some ambiguity / confusion related to a test profile having the name Patient (which conflicts with a resource name), so I updated the fixture and tests to reflect a unique name, and added a new fixture for a profile named Patient to test that name clash condition specifically.

* Simplify standalone instance names even if an inline instance has same id

* Update src/optimizer/plugins/SimplifyInstanceNameOptimizer.ts

Co-authored-by: Mint Thompson <mathompson@mitre.org>

* Properly handle leading _ in FHIR paths

Since FSH paths don't require a leading _, remove it from the common functions that extract FSH paths from FHIR paths.

---------

Co-authored-by: Mint Thompson <mathompson@mitre.org>