Special handling for CodeableReference type constraints
CodeableReference elements have to be handled a bit carefully because
they are sort of like Reference elements, but the FSH Reference keyword
usually tries to make things a Reference type. This can lead to issues
when a choice element has a CodeableReference type. Make separate rules
for constraining the type to CodeableReference and for constraining the
target profiles for CodeableReference.