Load the same automatic dependencies as SUSHI
Use SUSHI's function for loading automatic dependencies. There is a
minor discrepancy in the expected type, but the FHIR definitions are
only passed in so they can be passed to fhir-package-loader's
mergeDependency function. So, the ts-ignore here should be safe. Add a
note indicating the reason for the ts-ignore and the conditions under
which it can be removed.