ValueSet concept caret rules include system in path array
The SUSHI dependency is updated to 3.6.0, which supports including the
system in the path array. This is needed for rules on ValueSets. Add the
system to the path array when extracting caret rules on ValueSet concept
components. Caret rules on CodeSystem concepts don't need a system, but
still need the leading # as a separator.

Add optimizer that reorders rules on ValueSets so that concept rules are
grouped with their corresponding caret rules.

Update other tests for Invariants, ValueSets, and ConceptRules based on
other changes included in SUSHI 3.6.0.