Resolve URLs in ValueSet caret rule concept paths
Resolve URLs on caret rules after resolving URLs on ValueSet components
so that the caret rules will be able to use aliases that were created
when resolving the components.

Create new arrays to assign to path array when extracting caret rules on
concepts. This avoids the possibility of inadvertantly modifying
multiple rules with a single operation that mutates the path array.