Support code caret rules in ValueSets (#249)
* Support code caret rules in ValueSets

A ValueSet can use caret rules with a code path to set values on
included or excluded concepts. The typical use of this is to set a
designation, but other elements may also appear. Elements outside of a
concept are still unsupported and require the creation of an Instance.

* update jest-extended dependency to avoid downstream problems

* ValueSet concept caret rules include system in path array

The SUSHI dependency is updated to 3.6.0, which supports including the
system in the path array. This is needed for rules on ValueSets. Add the
system to the path array when extracting caret rules on ValueSet concept
components. Caret rules on CodeSystem concepts don't need a system, but
still need the leading # as a separator.

Add optimizer that reorders rules on ValueSets so that concept rules are
grouped with their corresponding caret rules.

Update other tests for Invariants, ValueSets, and ConceptRules based on
other changes included in SUSHI 3.6.0.

* Resolve URLs in ValueSet caret rule concept paths

Resolve URLs on caret rules after resolving URLs on ValueSet components
so that the caret rules will be able to use aliases that were created
when resolving the components.

Create new arrays to assign to path array when extracting caret rules on
concepts. This avoids the possibility of inadvertantly modifying
multiple rules with a single operation that mutates the path array.

* Add concepts without caret rules to optimizer test