Extract rules for Invariants. Use keyword for Invariant description, rules for everything else (#251)
* Add rules to Invariants

When extracting an Invariant, create assignment rules for properties
that are not represented using keywords. Use the same functions as the
InstanceProcessor for this. Add the paths for these elements to the
returned list of paths so that caret rules will not be created for these
elements.

* Use keyword for description, rules for everything else

When producing FSH for an Invariant, use the Description keyword for the
description element. For all other metadata elements, use assignment
rules instead. These rules will always come before any other rules on
the Invariant.

* Remove metadataToFSH override from ExportableInvariant