Process ValueSet that uses components from a contained system
When a ValueSet component references a contained system, the
valueset-system extension should be present at compose.system.extension.
When it is present, it is safe to throw it out, since SUSHI will add it
back in automatically. When it is removed, also remove the extension
list and containing underscore property if those become empty. This will
often allow the input to be processed as a ValueSet.