Use new FPL API
The FPL version 2 package includes API changes. Update classes,
functions, and tests that include loaded dependencies to use the new
API. Most of this is acquired through SUSHI's FHIRDefinitions class,
which extends FPL's BasePackageLoader class. The main thing to watch out
for is that FHIRDefinitions must be initialized before use, which is an
async operation.

The LakeOfFHIR class is updated to have an instance of FHIRDefinitions
for access to its fishing methods. Therefore, it also needs to be
initialized asynchronously. There may be some possible redesigns of the
LakeOfFHIR class in order to more fully make use of the new FPL API.