Use most recent SUSHI and FPL dependencies
For LakeOfFHIR, since assignMissingIds and removeDuplicateDefinitions
are preparatory steps, add them to the beginning of prepareDefs. Since
they should only be used as these preparatory steps, the methods are
also marked as private.

FSH grammar supports designation on a concept in a CodeSystem. These
test fixtures are therefore now supported and renamed accordingly. This
means that since only a missing name and id makes a CodeSystem
unprocessable, and an id will be added by LakeOfFHIR if it is missing,
it means that any CodeSystem resource is processable after
LakeOfFHIR.prepareDefs is called.

Add helper function to simplify test setup in
ResolveValueSetComponentRuleURLsOptimizer.

Mark FHIRProcessor tests as async, since the calls to restokeLake should
be awaited.

Minor cleanup in some other files.