Use new FPL API (#277)
* Use new FPL API

The FPL version 2 package includes API changes. Update classes,
functions, and tests that include loaded dependencies to use the new
API. Most of this is acquired through SUSHI's FHIRDefinitions class,
which extends FPL's BasePackageLoader class. The main thing to watch out
for is that FHIRDefinitions must be initialized before use, which is an
async operation.

The LakeOfFHIR class is updated to have an instance of FHIRDefinitions
for access to its fishing methods. Therefore, it also needs to be
initialized asynchronously. There may be some possible redesigns of the
LakeOfFHIR class in order to more fully make use of the new FPL API.

* Use most recent SUSHI and FPL dependencies

For LakeOfFHIR, since assignMissingIds and removeDuplicateDefinitions
are preparatory steps, add them to the beginning of prepareDefs. Since
they should only be used as these preparatory steps, the methods are
also marked as private.

FSH grammar supports designation on a concept in a CodeSystem. These
test fixtures are therefore now supported and renamed accordingly. This
means that since only a missing name and id makes a CodeSystem
unprocessable, and an id will be added by LakeOfFHIR if it is missing,
it means that any CodeSystem resource is processable after
LakeOfFHIR.prepareDefs is called.

Add helper function to simplify test setup in
ResolveValueSetComponentRuleURLsOptimizer.

Mark FHIRProcessor tests as async, since the calls to restokeLake should
be awaited.

Minor cleanup in some other files.

* Override SUSHI logger with GoFSH logger in FHIRDefinitions

Remove unneeded comments from LakeOfFHIR.

* Suppress logger output in CardRuleExtractor tests